pyproject.toml
[tool.poetry]
name = "pytest-bdd-ui-automation"
version = "0.1.0"
description = ""
authors = ["Lumunix <lumunix@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^7.1.2"
selenium="^4.4.3"
webdriver-manager="^3.8.3"
pytest-bdd = "6.1.1"
sttable = "*"
allure-pytest-bdd = "^2.13.1"


[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
